+++ packages/astro/src/cli/index.ts
@@ -1,7 +1,6 @@
 /* eslint-disable no-console */
 import fs from 'fs';
 import * as colors from 'kleur/colors';
-import { arch, platform } from 'node:os';
 import type { Arguments as Flags } from 'yargs-parser';
 import yargs from 'yargs-parser';
 import { ZodError } from 'zod';
@@ -73,57 +72,6 @@ async function printVersion() {
 	console.log(`  ${colors.bgGreen(colors.black(` astro `))} ${colors.green(`v${ASTRO_VERSION}`)}`);
 }
 
-async function printInfo({
-	cwd,
-	flags,
-	logging,
-}: {
-	cwd?: string;
-	flags?: Flags;
-	logging: LogOptions;
-}) {
-	const whichPm = await import('which-pm');
-	const packageManager = await whichPm.default(process.cwd());
-	let adapter = "Couldn't determine.";
-	let integrations = [];
-
-	const MAX_PADDING = 25;
-	function printRow(label: string, value: string) {
-		const padding = MAX_PADDING - label.length;
-		console.log(`${colors.bold(label)}` + ' '.repeat(padding) + `${colors.green(value)}`);
-	}
-
-	try {
-		const { userConfig } = await openConfig({
-			cwd,
-			flags,
-			cmd: 'info',
-			logging,
-		});
-		if (userConfig.adapter?.name) {
-			adapter = userConfig.adapter.name;
-		}
-		if (userConfig.integrations) {
-			integrations = (userConfig?.integrations ?? [])
-				.filter(Boolean)
-				.flat()
-				.map((i: any) => i?.name);
-		}
-	} catch (_e) {}
-	console.log();
-	const packageManagerName = packageManager?.name ?? "Couldn't determine.";
-	printRow('Astro version', `v${ASTRO_VERSION}`);
-	printRow('Package manager', packageManagerName);
-	printRow('Platform', platform());
-	printRow('Architecture', arch());
-	printRow('Adapter', adapter);
-	let integrationsString = "None or couldn't determine.";
-	if (integrations.length > 0) {
-		integrationsString = integrations.join(', ');
-	}
-	printRow('Integrations', integrationsString);
-}
-
 /** Determine which command the user requested */
 function resolveCommand(flags: Arguments): CLICommand {
 	const cmd = flags._[2] as string;
@@ -183,6 +131,7 @@ async function runCommand(cmd: string, flags: yargs.Arguments) {
 			await printVersion();
 			return process.exit(0);
 		case 'info': {
+			const { printInfo } = await import('./info/index.js');
 			await printInfo({ cwd: root, flags, logging });
 			return process.exit(0);
 		}
